added check on clear to not delete locked tests

+++ core/php/removeAllTmpTests.php
@@ -4,7 +4,7 @@ $files = array_diff(scandir($dir), array('..', '.'));
 foreach($files as $key => $value)
 {
 	$path = realpath($dir.DIRECTORY_SEPARATOR.$value);
-    if(is_file($path))
+    if(is_file($path) && strpos($path, "LOCK") === -1)
     {
     	unlink($path);
     }
